requirements.txt
pandas
plotly
streamlit
pillow
openpyxl
pyxlsb